// LaptopShop.ClientApp/src/app/features/cart/model/cart-model.ts
export interface CartItem {
  productID: number;
  quantity: number;
  productName: string;
  productPrice: number;
  imageURL: string;
}

// export interface Product {
//   productID: number;
//   name: string;
//   price: number;
//   solanxem: number;
//   stockQuantity: number;
//   description: string;
//   imageURL: string;
//   createdDate: string;
//   brandID: number;
// }